Move create collective garphql v2 to component
