+++ lib/graphql/mutations.js
@@ -2,8 +2,6 @@ import { graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 import { pick, isArray } from 'lodash';
 
-const gqlV2 = gql; // Needed for lint validation of api v2 schema.
-
 import { getLoggedInUserQuery, getCollectiveToEditQueryFields } from './queries';
 
 export const createUserQuery = gql`
@@ -32,22 +30,6 @@ const createCollectiveQuery = gql`
       website
       twitterHandle
       isIncognito
-    }
-  }
-`;
-
-const createCollectiveQueryV2 = gqlV2`
-  mutation createCollective(
-    $collective: CreateCollectiveInput!
-    $host: AccountInput
-    $automateApprovalWithGithub: Boolean
-  ) {
-    createCollective(collective: $collective, host: $host, automateApprovalWithGithub: $automateApprovalWithGithub) {
-      name
-      slug
-      tags
-      description
-      githubHandle
     }
   }
 `;
@@ -164,39 +146,6 @@ export const addCreateCollectiveMutation = graphql(createCollectiveQuery, {
           data.LoggedInUser.memberOf.push({
             __typename: 'Member',
             collective: createCollective,
-            role: 'ADMIN',
-          });
-          store.writeQuery({ query: getLoggedInUserQuery, data });
-        },
-      });
-    },
-  }),
-});
-
-export const addCreateCollectiveMutationV2 = graphql(createCollectiveQueryV2, {
-  options: {
-    context: { apiVersion: '2' },
-  },
-  props: ({ mutate }) => ({
-    createCollectiveV2: async ({ collective, host }) => {
-      const CreateCollectiveInputType = pick(collective, [
-        'slug',
-        'name',
-        'description',
-        'githubHandle',
-        'tags',
-        CreateCollectiveInputType,
-      ]);
-      return await mutate({
-        variables: {
-          collective: CreateCollectiveInputType,
-          host: host,
-        },
-        update: (store, { data: { createCollectiveV2 } }) => {
-          const data = store.readQuery({ query: getLoggedInUserQuery });
-          data.LoggedInUser.memberOf.push({
-            __typename: 'Member',
-            collective: createCollectiveV2,
             role: 'ADMIN',
           });
           store.writeQuery({ query: getLoggedInUserQuery, data });
